% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archived_ads.R
\name{aa_process_InsightsRangeValue}
\alias{aa_process_InsightsRangeValue}
\title{Flatten an InsightsRangeValue}
\usage{
aa_process_InsightsRangeValue(l, field_name)
}
\arguments{
\item{l}{a named list as returned by the ads library API}

\item{field_name}{name of the InsightsRangeValue field contained in l}
}
\value{
a list with the `field_name` field replaced with upper and lower
  bounds
}
\description{
Flatten an InsightsRangeValue
}
\details{
See https://developers.facebook.com/docs/graph-api/reference/insights-range-value/
for documentation on the InsightsRangeValue data type
}
