% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\name{ad_row}
\alias{ad_row}
\title{Create a single row in the ad table}
\usage{
ad_row(row)
}
\arguments{
\item{row}{a single row in the response}
}
\value{
tibble with a single row
}
\description{
Create a single row in the ad table
}
