% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\name{ad_table}
\alias{ad_table}
\title{Create an ad_table from results}
\usage{
ad_table(results, handle_dates = TRUE)
}
\arguments{
\item{results}{data from an ads_archive response}

\item{handle_dates}{if true, convert dates columns to date}
}
\value{
dataframe with one row per ad
}
\description{
Create an ad_table from results
}
