% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_check_for_token}
\alias{adlib_check_for_token}
\title{Check whether a token has been set.}
\usage{
adlib_check_for_token(path = DEFAULT_TOKEN_PATH)
}
\arguments{
\item{path}{path to check for token}
}
\value{
TRUE if there is a token
}
\description{
Check whether a token has been set.
}
\examples{
adlib_check_for_token
}
