% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib_auth.R
\name{adlib_config_init}
\alias{adlib_config_init}
\title{Initialize the configuration file.}
\usage{
adlib_config_init()
}
\value{
True
}
\description{
This is where you tell adslibrary the details about your app
so that it can fetch long-term access tokens for you. You'll need:
}
\details{
- Your Application ID
- Your App secret

These are both located in your app's Basic Settings, which you can find
by signing into developers.facebook.com.
}
\examples{
\dontrun{
adlib_config_init()
}
}
