% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_data_frame}
\alias{adlib_data_frame}
\title{Convert Ad Library response to tibble}
\usage{
adlib_data_frame(response)
}
\arguments{
\item{response}{a response form adlib_get}
}
\value{
tibble
}
\description{
Convert Ad Library response to tibble
}
