% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_get}
\alias{adlib_get}
\title{Get results from Facebook ads library}
\usage{
adlib_get(params, token = adlib_read_token(DEFAULT_TOKEN_PATH))
}
\arguments{
\item{params}{a list of parameters build with adlib_build_params}

\item{token}{an access token from Facebook}
}
\value{
the raw response from Facebook Ads library
}
\description{
Get results from Facebook ads library
}
