% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_get_paginated}
\alias{adlib_get_paginated}
\title{Paginated query to the ad library API}
\usage{
adlib_get_paginated(query, max_gets = 100, token = token_get())
}
\arguments{
\item{query}{an Ad Library query}

\item{max_gets}{The maximum number of calls to adlib_get that this will
generate}

\item{token}{an access_token}
}
\value{
a paginated_adlib_data_response object. The response has the
following attributes.
\itemize{
  \item responses - The list of individual responses
  \item has_next - A boolean value indicating whether there are still
  pages
  \item next_page - The URL to call the next page of responses
}
}
\description{
Paginated query to the ad library API
}
