% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_link}
\alias{adlib_link}
\title{Generate a URL to view an ad in the ad library}
\usage{
adlib_link(ad_id)
}
\arguments{
\item{ad_id}{a vector of ad IDs}
}
\value{
a vector of URLs
}
\description{
This is a different and more comprehensive view of the ad from what is
provided in the snapshot link that is returned by the API.
}
\examples{
ids <- c("399107662471074", "313132487749670")
adlib_link(ids)
}
