% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_read_token}
\alias{adlib_read_token}
\title{Read the token}
\usage{
adlib_read_token(path_to_token = check_for)
}
\arguments{
\item{path_to_token}{path to the token file}
}
\value{
returns True
}
\description{
Read the token
}
