% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_set_token}
\alias{adlib_set_token}
\title{Set Token}
\usage{
adlib_set_token(save_as = DEFAULT_TOKEN_PATH)
}
\arguments{
\item{save_as}{path to token file.}
}
\value{
Returns TRUE
}
\description{
Sets the access token from Facebook. Go to https://developers.facebook.com/tools/explorer/
to find your token, run adlib_set_token(), and paste your token at the prompt.
}
\examples{
\dontrun{
adlib_set_token()
}
}
