% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{adlib_setup}
\alias{adlib_setup}
\title{Set up application ID and app secret.}
\usage{
adlib_setup()
}
\description{
This lets you store your application ID and app secret securely
in your system's credential store (using the keyring package).
}
\details{
The application ID and app secret are used to obtain a long-term
access token.
}
\seealso{
\code{\link{adlib_set_longterm_token}} for obtaining a
long-term access token.
}
