% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_supported_fields}
\alias{adlib_supported_fields}
\title{Supported fields}
\usage{
adlib_supported_fields()
}
\value{
a character vector of fields
}
\description{
These are the fields that are fully supported by this package.
}
