% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{adlib_update_token}
\alias{adlib_update_token}
\title{Obtain a long-term token.}
\usage{
adlib_update_token()
}
\value{
TRUE
}
\description{
This function exchanges a short-term token for a long-term token.
The long-term token will expire in about 60 days. The token is
written to your .adlib_config.yml file.
}
