% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\name{as_tibble.adlib_data_response}
\alias{as_tibble.adlib_data_response}
\title{Convert a data response to tibble}
\usage{
\method{as_tibble}{adlib_data_response}(x, censor_access_token = NULL, ...)
}
\arguments{
\item{x}{an adlib_data_response object returned by adlib_get}

\item{censor_access_token}{should access tokens be censored?}

\item{...}{additional arguments to pass to conversion function}
}
\value{
a tibble
}
\description{
Convert a data response to tibble
}
