% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_clear_token}
\alias{config_clear_token}
\title{Remove the token from your config file.}
\usage{
config_clear_token()
}
\value{
TRUE
}
\description{
Your access token can provide full access to your Facebook account. For this reason, you
may wish to remove it when you're finished working. Use this file to delete your access
token from your config file.
}
