% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\name{demographic_row}
\alias{demographic_row}
\title{Result Row}
\usage{
demographic_row(result_row)
}
\arguments{
\item{result_row}{demographic data for a single ad}
}
\value{
a dateframe with one row
}
\description{
Result Row
}
