% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\name{demographic_table}
\alias{demographic_table}
\title{Turn data from the data field in response content into a demographics table}
\usage{
demographic_table(results)
}
\arguments{
\item{results}{the data field of a valid response}
}
\value{
a dataframe
}
\description{
Turn data from the data field in response content into a demographics table
}
