% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{extract_error_message}
\alias{extract_error_message}
\title{Extract an error message from a response to pass as an R error}
\usage{
extract_error_message(response)
}
\arguments{
\item{response}{a response from the Ad Library API}
}
\value{
a character vector of length 1 containing an error message
}
\description{
Extract an error message from a response to pass as an R error
}
