% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{format_array}
\alias{format_array}
\title{Format a vector as a json array}
\usage{
format_array(items)
}
\arguments{
\item{items}{a vector of items to format}
}
\value{
a string
}
\description{
Format a vector as a json array
}
