% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_api.R
\name{graph_api_endpoint}
\alias{graph_api_endpoint}
\title{API endpoints}
\usage{
graph_api_endpoint(api = c("ads_archive", "access_token"),
  version = "v5.0")
}
\arguments{
\item{api}{which API to return the end point for. Currently two endpoints are supported:
"ads_archive", for querying the Ad Library, and "access_token", for requesting
long-term access tokens.}

\item{version}{the version of the API to use.}
}
\value{
the endpoint URL
}
\description{
API endpoints
}
