% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_api.R
\name{graph_get}
\alias{graph_get}
\title{Get results from Facebook ads library}
\usage{
graph_get(end_point, params, token = token_current()[["token"]])
}
\arguments{
\item{end_point}{the graph API endpoint to send a request to}

\item{params}{a list of parameters build with adlib_build_params}

\item{token}{an access token from Facebook}
}
\value{
the raw response from Facebook Ads library
}
\description{
Get results from Facebook ads library
}
