% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\name{n_responses}
\alias{n_responses}
\title{Number of pages in a paginated response}
\usage{
n_responses(x)
}
\arguments{
\item{x}{a paginated data response}
}
\description{
Number of pages in a paginated response
}
