% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radlibrary-package.R
\docType{package}
\name{Radlibrary-package}
\alias{Radlibrary}
\alias{Radlibrary-package}
\title{Radlibrary: Facebook Ad Library API Tool}
\description{
Query the Facebook Ads Library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://facebookresearch.github.io/Radlibrary/}
  \item Report bugs at \url{https://github.com/facebookresearch/Radlibrary/issues}
}

}
\author{
\strong{Maintainer}: Colin Fraser \email{colinfraser@fb.com}

Authors:
\itemize{
  \item Kyle Shank \email{kylescotshank@fb.com}
}

}
\keyword{internal}
