% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_build_query}
\alias{adlib_build_query}
\title{Build an Ad Library Query}
\usage{
adlib_build_query(ad_reached_countries, ad_active_status = c("ACTIVE",
  "INACTIVE", "ALL"), ad_type = c("POLITICAL_AND_ISSUE_ADS",
  "HOUSING_ADS", "NEWS_ADS", "UNCATEGORIZED_ADS", "ALL"), bylines = NULL,
  impression_condition = c("HAS_IMPRESSIONS_LIFETIME",
  "HAS_IMPRESSIONS_YESTERDAY", "HAS_IMPRESSIONS_LAST_7_DAYS",
  "HAS_IMPRESSIONS_LAST_30_DAYS", "HAS_IMPRESSIONS_LAST_90_DAYS"),
  publisher_platform = "FACEBOOK", search_page_ids = NULL,
  search_terms = NULL)
}
\arguments{
\item{ad_reached_countries}{Vector of ISO country codes. Facebook delivered the ads in these countries.}

\item{ad_active_status}{Status of the ads. One of 'ACTIVE', 'INACTIVE', or 'ALL'.}

\item{ad_type}{The type of ad. One of 'ALL', 'POLITICAL_AND_ISSUE_ADS', 'HOUSING_ADS', 'NEWS_ADS', or 'UNCATEGORIZED_ADS'. Currently only 'POLITICAL_AND_ISSUE_ADS' is supported.}

\item{bylines}{Filter results for ads with a paid for by disclaimer byline, such as political ads that reference “immigration” paid for by “ACLU”.}

\item{impression_condition}{When the ad most recently had impressions. One of "HAS_IMPRESSIONS_LIFETIME", "HAS_IMPRESSIONS_YESTERDAY", "HAS_IMPRESSIONS_LAST_7_DAYS", "HAS_IMPRESSIONS_LAST_30_DAYS", "HAS_IMPRESSIONS_LAST_90_DAYS"}

\item{publisher_platform}{The platform on which the ads appeared. One or more of "FACEBOOK", "INSTAGRAM", "AUDIENCE_NETWORK", "MESSENGER", "WHATSAPP".}

\item{search_page_ids}{A vector of up to 10 page IDs to search.}

\item{search_terms}{A search string.}
}
\value{
A list of params
}
\description{
This will build a list that can be passed to adlib_get.
Visit https://developers.facebook.com/docs/marketing-api/reference/ads_archive/
for full API documentation.
}
