% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{adlib_config_setup}
\alias{adlib_config_setup}
\title{Initialize the configuration file.}
\usage{
adlib_config_setup()
}
\value{
True
}
\description{
This is where you tell adslibrary the details about your app
so that it can fetch long-term access tokens for you. You'll need:
}
\details{
- Your Application ID
- Your App secret

These are both located in your app's Basic Settings, which you can find
by signing into developers.facebook.com.

This will save a file called .fb_adlib_config.yaml to your home directory
containing your App ID, App secret, and eventually a token.
}
\examples{
\dontrun{
adlib_config_setup()

}

}
