% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlib.R
\name{adlib_get}
\alias{adlib_get}
\title{Query the Ad Library API}
\usage{
adlib_get(params, token = token_get())
}
\arguments{
\item{params}{a list of parameters.}

\item{token}{an access_token.}
}
\value{
an http response
}
\description{
The `params` argument should be a list that is built with adlib_build_query().
The `token` argument can be a short-term token pasted from https://developers.facebook.com/tools/explorer/,
or a long-term token stored in your password store.
}
