% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{adlib_set_longterm_token}
\alias{adlib_set_longterm_token}
\title{Set long term token}
\usage{
adlib_set_longterm_token()
}
\description{
Once you've entered your Application ID and App secret, you can
obtain a long-term token that will expire in about 60 days. This will be stored
in your operating system's secrets manager.
}
\seealso{
\code{\link{adlib_setup}} for storing your App ID and App Secret.
}
