% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data.R
\name{as_tibble.adlib_data_response}
\alias{as_tibble.adlib_data_response}
\title{Convert a data response to tibble}
\usage{
\method{as_tibble}{adlib_data_response}(response, type = c("ad", "demographic", "region"), ...)
}
\arguments{
\item{response}{an adlib_data_response object returned by adlib_get}

\item{type}{the type of table to output. One of c("ad", "demographic",
"region").}

\item{...}{additional arguments to pass to conversion function}
}
\value{
a tibble
}
\description{
Convert a data response to tibble
}
