% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_api.R
\name{graph_get}
\alias{graph_get}
\title{Get results from Facebook Ad library}
\usage{
graph_get(service, params, token = token_get()[["token"]])
}
\arguments{
\item{service}{the graph API endpoint to send a request to}

\item{params}{a list of parameters build with adlib_build_params}

\item{token}{an access token from Facebook}
}
\value{
the raw response from Facebook Ad library
}
\description{
Get results from Facebook Ad library
}
